Rename scripts/__init__.py to cli.py
